move tests to new class
